requirements.txt
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
contourpy==1.1.1
cycler==0.11.0
dataclasses==0.6
datasets==2.14.5
diffusers @ git+https://github.com/huggingface/diffusers@9cfd4ef0746076febb589788ec47df9e2db43d65
dill==0.3.7
docker-pycreds==0.4.0
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.37
GPUtil==1.4.0
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirerpy==0.3.4
Jinja2==3.1.2
kiwisolver==1.4.5
lightning-utilities==0.9.0
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
pfzy==0.3.4
Pillow==10.0.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scipy==1.10.1
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torch-fidelity==0.3.0
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wandb==0.15.11
wcwidth==0.2.8
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
